Procedimentos de criação, deleção e update de instituições criado

+++ src/main/java/com/orizon/webdriver/domain/service/InstitutionServiceImpl.java
@@ -1,5 +1,6 @@
 package com.orizon.webdriver.domain.service;
 
+import com.orizon.webdriver.domain.exceptions.ENFieldException;
 import com.orizon.webdriver.domain.exceptions.InvalidInstitutionException;
 import com.orizon.webdriver.domain.model.Institution;
 import com.orizon.webdriver.infra.repositories.InstitutionRepository;
@@ -7,6 +8,8 @@ import com.orizon.webdriver.domain.ports.service.InstitutionService;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Objects;
 
 @Service
 @Transactional
@@ -30,17 +33,35 @@ public class InstitutionServiceImpl implements InstitutionService {
     }
 
     @Override
-    public void save(Institution institution){
+    public void create(String name, String socialCause){
+        Objects.requireNonNull(name, () -> {throw new ENFieldException();});
+        Objects.requireNonNull(socialCause, () -> {throw new ENFieldException();});
+        Institution institution = new Institution(name, socialCause);
         institutionDAO.save(institution);
     }
 
     @Override
     public void delete(Long id) {
-        institutionDAO.deleteById(id);
+        Objects.requireNonNull(id, () -> {throw new ENFieldException();});
+        if(findOne(id) != null){
+            institutionDAO.deleteById(id);
+        } else {
+            throw new InvalidInstitutionException();
+        }
     }
 
     @Override
-    public void update(Institution institution) {
+    public void update(Long id, String name, String socialCause) {
+        Institution institution = findOne(id);
+        if(name == null && socialCause == null){
+            return;
+        }
+        if(name == null && !socialCause.isBlank()){
+            institution.setSocialCause(socialCause);
+        }
+        if (socialCause == null && !name.isBlank()){
+            institution.setName(name);
+        }
         institutionDAO.save(institution);
     }
 }
